Update codebase for Emacs 29.x
The Common Lisp support changed in Emacs over time.  These are
machanical changes to accomodate these changes.

All ERT test in kv-tests.el pass with these changes in Emacs 29.0.50

The following warnings are fixed:

Warning: Package cl is deprecated
Warning: docstring has wrong usage of unescaped single quotes (use \= or different quoting)
Warning: ‘defun*’ is an obsolete alias (as of 27.1); use ‘cl-defun’ instead.
Warning: ‘destructuring-bind’ is an obsolete alias (as of 27.1); use ‘cl-destructuring-bind’ instead.
Warning: ‘flet’ is an obsolete macro (as of 24.3); use either ‘cl-flet’ or ‘cl-letf’.
Warning: ‘loop’ is an obsolete alias (as of 27.1); use ‘cl-loop’ instead.

This resolves #11